pass first stage

+++ src/Server.cpp
@@ -319,9 +319,9 @@ void printTableLeafPage(std::vector<char> &database_file , unsigned short num_ta
     // Calculate factorialass b
     for (int i = 0; i < num_table; i++) {
         int startIndex = (2*i)+start; 
-        unsigned short byte1 = static_cast<unsigned char>(database_file[startIndex]);
-        unsigned short byte2 = static_cast<unsigned char>(database_file[startIndex+1]);
-        unsigned short page_address =   (byte1 << 8) | (byte2);
+        unsigned long byte1 = static_cast<unsigned char>(database_file[startIndex]);
+        unsigned long byte2 = static_cast<unsigned char>(database_file[startIndex+1]);
+        unsigned long page_address =   (byte1 << 8) | (byte2);
         cellAddress.push_back(pageStart+page_address);
         // std::cerr << "Logs from your program will appeaar here" << i <<"  "<<page_address+pageStart<<std::endl;
          if (printTables || debugStage ){
